requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==23.2.0
autopep8==2.3.1
billiard==4.2.1
black==24.4.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-schema-adapter==3.0.6
environ==1.0
idna==3.10
inflection==0.5.1
Inflector==3.0.1
kombu==5.5.0
Markdown==3.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
